% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlookr.R
\docType{package}
\name{kodlookr-package}
\alias{kodlookr}
\alias{kodlookr-package}
\title{kodlookr: Tdkookr 패키지의 문서 한글화 패키지}
\description{
ddlookr 패키지는 데이터 분석 과정에서의 데이터 진단, 데이터 탐색(EDA) 및
변수의 변환 기능을 제공합니다.
}
\details{
dlookr는 다음과 같은 세 개의 목적이 있습니다.:

\itemize{
\item 분석할 데이터를 획득하면 먼저, 데이터에 포함된 오류 여부 등을 판단하여
데이터 분석이 가능한 수준인지  검토하는 데이터 진단을 수행합니다. 이를 통해서
수정하거나 제거할 변수를 가려밸 수 있습니다.
\item 탐색적 데이터분석(EDA) 프로세스를 통해서 데이터 분포를 이해합니다.
이 과정에서 예측 모델을 위한 목적 변수(target variable)와 예측변수
(predictor variables) 간의 관계를 이해할 수 있습니다.
\item 누락된 값(결측치, missing values)과 이상값(outliers)을 대체하고,
연속형 변수를 표준화하거나 치우친 분포의 변수를 정규분포화합니다.
또한, 연속형 변수를 범주형 변수로 변환하는 비닝을 쉽게 수행할 수 있습니다.
}

dlookr 패키지의 사용 방법을 자세하게 이해하기 위해서는 비네트(vignettes)부터 시작하세요.:
\code{browseVignettes(package = "kodlookr")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bit2r/kodlookr}
  \item Report bugs at \url{https://github.com/bit2r/kodlookr/issues}
}

}
\author{
\strong{Maintainer}: Choonghyun Ryu \email{choonghyun.ryu@gmail.com}

Authors:
\itemize{
  \item Gildong Hong
}

}
