% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{plot.bins}
\alias{plot.bins}
\title{"bins" 객체의 분포 시각화}
\usage{
## 'bins' 클래스 대상의 S3 메서드
plot(x, typographic = TRUE, ...)
}
\arguments{
\item{x}{"bins" 클래스의 객체로 binning()을 호출했을 때의 일반적인 결과입니다.}

\item{typographic}{logical. 적용 여부는 ggplot2 시각화에서 타이포 그래픽 요소의
중점에 따라 변경됩니다.
기본값은 TRUE 입니다. 값이 TRUE이면 hrbrthemes 패키지를 사용한 타이포 그래픽 요소에
중점을 둔 기본 테마를 제공합니다.}

\item{...}{그래픽 매개변수와 같이 메소드에 전달되는 인수(파라미터 참고)}
}
\description{
단일 화면에서 두 개의 플롯을 시각화합니다.
상단의 플롯은 수준의 빈도를 나타내는 히스토그램입니다.
하단의 플롯은 수준의 빈도를 나타내는 막대 차트입니다.
}
\examples{
# 예제를 위한 데이터 생성
heartfailure2 <- heartfailure
heartfailure2[sample(seq(NROW(heartfailure2)), 20), "platelets"] <- NA

# type 인수의 기본값인  "quantile"을 이용한 platelets 변수의 비닝
bin <- binning(heartfailure2$platelets, nbins = 5)
plot(bin)

# 다른 타입의 인수 사용
bin <- binning(heartfailure2$platelets, nbins = 5, type = "equal")
plot(bin)

bin <- binning(heartfailure2$platelets, nbins = 5, type = "pretty")
plot(bin)

bin <- binning(heartfailure2$platelets, nbins = 5, type = "kmeans")
plot(bin)

bin <- binning(heartfailure2$platelets, nbins = 5, type = "bclust")
plot(bin)

}
\seealso{
\code{\link{binning}}, \code{\link{print.bins}}, \code{\link{summary.bins}}.
}
