% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{summary.bins}
\alias{summary.bins}
\alias{print.bins}
\title{비닝된 변수의 요약}
\usage{
## 'bins' 클래스 대상의 S3 메서드
summary(object, ...)

## 'bins' 클래스 대상의 S3 메서드
print(object, ...)
}
\arguments{
\item{object}{"bins", "optimal_bins" 객체로,
binning()을 호출했을 때의 일반적인 결과}

\item{x}{"bins"와 "optimal_bins" 클래스 객체로,
binning()을 호출했을 때의 일반적인 결과}

\item{...}{다른 메서드에서 또는 다른 메서드로 전달되는 추가 인수}
}
\value{
summary.bins() 함수는 객체에 제공된 비닝된 요약 통계를 계산하고
data.frame 형태로 반환합니다. 데이터 프레임의 변수는 다음과 같습니다.
\itemize{
\item levels : 요인의 수준
\item freq : 요인의 빈도
\item rate : 수준의 상대 빈도이며, 백분율이 아닙니다.
}
}
\description{
"bins", "optimal_bins"을 대상으로 한 요약 방법
}
\details{
print.bins()는 "bins"와 "optimal_bins" 객체의 정보를 정확하게 출력합니다.
이것은 bins의 빈도, 비닝의 방법, bins의 개수를 포함합니다.
summary.bins()는 각 bins의 수준 빈도와 상대 빈도를 포함한 data.frame를 반환합니다.

비닝의 이해를 위해서 vignette("transformation") 명령어로 비네트를 참고하세요.
}
\examples{
# 예제를 위한 데이터 생성
heartfailure2 <- heartfailure
heartfailure2[sample(seq(NROW(heartfailure2)), 20), "platelets"] <- NA

# type 인수의 기본값인 "quantile"을 이용한 platelets 변수의 비닝
bin <- binning(heartfailure2$platelets)

# bins 클래스 객체의 출력
bin

# bins 클래스 객체의 요약
summary(bin)
}
\seealso{
\code{\link{binning}}
}
