% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{summary.performance_bin}
\alias{summary.performance_bin}
\title{비닝된 변수의 성능 요약}
\usage{
## 'performance_bin' 클래스 대상의 S3 메서드
summary(object, ...)
}
\arguments{
\item{object}{"performance_bin" 클래스 객체로,
performance_bin()을 호출했을 때의 일반적인 결과}

\item{...}{다른 메서드에서 또는 다른 메서드로 전달되는 추가 인수}
}
\value{
NULL.
}
\description{
"performance_bin"에 대한 요약 방법으로 이항 분류 모델의 성능을
평가하기 위한 요약 메트릭.
}
\details{
print()는 "performance_bin" 객체의 비닝 테이블 정보만을 출력합니다.
summary.performance_bin() 일반 메트릭 및 유의성 검정을 따른 결과를 포함합니다.:
\itemize{
\item Binning Table : 빈(bins)별 메트릭
\itemize{
\item CntRec, CntPos, CntNeg, RatePos, RateNeg, Odds, WoE, IV, JSD, AUC.
}
\item 일반 메트릭
\itemize{
\item 지니 계수
\item 제프리(Jeffrey)의 정보 가치
\item 젠슨-샤논(Jensen-Shannon) 발산
\item 콜모고로프-스미르노프(Kolmogorov-Smirnov) 통계량
\item 허핀달-허쉬만(Herfindahl-Hirschman) 지수
\item 정규화된 허핀달-허쉬만(Herfindahl-Hirschman) 지수
\item 크래머(Cramer)의 V 통계량
}
\item 유의성 검정 표
}
}
\examples{
# 예제를 위한 데이터 생성
heartfailure2 <- heartfailure

set.seed(123)
heartfailure2[sample(seq(NROW(heartfailure2)), 5), "creatinine"] <- NA

# 타깃 변수를 0(음)과 1(양)로 변경
heartfailure2$death_event_2 <- ifelse(heartfailure2$death_event \%in\% "Yes", 1, 0)

# creatinine에서 platelets_bin으로 비닝
breaks <- c(0,  1,  2, 10)
heartfailure2$creatinine_bin <- cut(heartfailure2$creatinine, breaks)

# 비닝된 변수의 성능 진단
perf <- performance_bin(heartfailure2$death_event_2, heartfailure2$creatinine_bin)
perf
summary(perf)

# plot(perf)

# 비닝된 변수의 성능 진단 (NA 제외)
perf <- performance_bin(heartfailure2$death_event_2, heartfailure2$creatinine_bin, na.rm = TRUE)
perf
summary(perf)

# plot(perf)

}
\seealso{
\code{\link{performance_bin}}, \code{\link{plot.performance_bin}}, \code{\link{binning_by}},
\code{\link{summary.optimal_bins}}.
}
